requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-pwa==1.0.10
django-rest-authtoken==2.1.4
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
Markdown==3.3.7
Pillow==9.2.0
python-decouple==3.6
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
zipp==3.8.1
